requirements.txt
Django==1.10.4
numpy==1.11.3
protobuf==3.1.0.post1
six==1.10.0
pillow==4.1.0